updated code for transaction management

+++ 2.0-SpringTransaction/src/main/java/more/pranav/springtransaction/repository/BusRepository.java
@@ -3,7 +3,12 @@ package more.pranav.springtransaction.repository;
 import more.pranav.springtransaction.entity.Bus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Repository
-public interface BusRepository extends JpaRepository<Bus, Long> {
+public interface BusRepository extends JpaRepository<Bus, Integer> {
+
 }
